Refactoring of the files: interface class to Constraint.h and Constraint_BodyToGroundPosition added
